requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
pydantic
pytesseract
pdf2image
Pillow
openai
python-dotenv
requests
pytest
httpx
langchain
